make NamespaceSelector its own component
